# apis/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/backups": {
            "get": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Возвращает список доступных резервных копий",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Backups"
                ],
                "summary": "Получение списка резервных копий",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/apis.BackupFileInfo"
                            }
                        }
                    },
                    "400": {
                        "description": "Failed to authenticate.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Создает новую резервную копию",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Backups"
                ],
                "summary": "Создание резервной копии",
                "parameters": [
                    {
                        "description": "Данные для создания резервной копии",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apis.BackupCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Failed to authenticate.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/backups/{key}": {
            "get": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Загружает резервную копию по указанному ключу",
                "tags": [
                    "Backups"
                ],
                "summary": "Загрузка резервной копии",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ключ резервной копии",
                        "name": "key",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Токен доступа",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Not exists.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Удаляет резервную копию по указанному ключу",
                "tags": [
                    "Backups"
                ],
                "summary": "Удаление резервной копии",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ключ резервной копии",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Failed to authenticate.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/backups/{key}/restore": {
            "post": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Запускает процесс восстановления резервной копии по указанному ключу",
                "tags": [
                    "Backups"
                ],
                "summary": "Восстановление резервной копии",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ключ резервной копии",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Failed to authenticate.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "apis.BackupCreateRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "apis.BackupFileInfo": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "modified": {
                    "type": "object"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
...
    }
}